Remove priority arbitrator
